AE-32

+++ backend/src/main/java/com/embraer/backend/item/services/ListItemsByChassi/ListItemsByChassiController.java
@@ -16,7 +16,7 @@ public class ListItemsByChassiController {
 
 	@CrossOrigin
 	@GetMapping
-	public ResponseEntity<ListItemsResponseDTO> listItemsByChassy(@PathVariable("chassi") Integer chassi) {
+	public ResponseEntity<ListItemsResponseDTO> listItemsByChassy(@PathVariable("chassi") Long chassi) {
 		
 		ListItemsResponseDTO listItems = itemsByChassiService.execute(chassi);
